requirements.txt
alembic==1.13.3
amqp==5.2.0
aniso8601==9.0.1
billiard==4.2.1
blinker==1.8.2
boto3==1.35.54
botocore==1.35.54
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
factory_boy==3.3.1
Faker==30.8.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-enum==1.5.1
packaging==24.1
password-strength==0.0.3.post2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.2.0
s3transfer==0.10.3
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
